Update main.asm
Added to the Ai victory and Player victory outcomes. Only thing left to fix is horizontal checker. Seems to not be able to tell if the piece is placed in the middle of a horizontal pattern, for example, if you have XX-X, and place the final X in between it will not be able to tell. There's also the case when you have XXX- and it doesn't detect that victory if you put the final X there. It works sometimes however.